(PC-36205) refactor(home): extract hook to queries
